Update XMLTV generator
