+++ resources/lib/__init__.py
@@ -1,4 +1,2 @@
 # -*- coding: utf-8 -*-
 # pylint: disable=missing-module-docstring
-from .m3u_generator import M3UGenerator
-from .xmltv_generator import XMLTVGenerator
